requirements.txt
fastapi
openai
python-dotenv
pydantic
langchain
faiss-cpu
tiktoken
watchdog
streamlit
google-cloud-bigquery
langchain-community
langchain-openai